+++ src/resources/index.ts
@@ -23,6 +23,7 @@ export * from "./candidate";
 export * from "./job";
 export * from "./client";
 export * from "./recruiter";
+export * from "./contact";
 export * from "./process";
 export * from "./resume";
 export * from "./attachment";